production/requirements.txt
dj-database-url==0.4.1
Django==2.2.0
django-extensions==1.6.1
django-filter==1.0.4
django-material==1.0.0
djangorestframework==3.7.7
gunicorn==19.6.0
psycopg2==2.7.5
python-env==1.0.0
whitenoise==3.2
django-cors-headers==3.7.0
Flask==1.1.2
requests==2.25.1